requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dm-tree==0.1.8
exceptiongroup==1.1.1
executing==2.0.1
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.49.0
frozenlist==1.3.3
fsspec==2024.2.0
gast==0.5.4
google-pasta==0.2.0
gradio==4.21.0
gradio_client==0.12.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.4
importlib_resources==6.1.3
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keras==3.0.5
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.4
namex==0.0.7
nest-asyncio==1.6.0
numpy==1.26.4
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.2.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyJWT==2.6.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
requests==2.28.2
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.2
selenium==4.8.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.36.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
trio==0.22.0
trio-websocket==0.10.2
twilio==8.0.0
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.15
uvicorn==0.28.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.8.2
